fix khs progress; changing profile routes, add ProfileController for all user; del some unused resources; etc
