+++ app/Http/Controllers/DosenWaliController.php
@@ -10,8 +10,6 @@ use App\Models\DosenWali;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Http\Request;
 use Maatwebsite\Excel\Facades\Excel;
-use Illuminate\Validation\Rule;
-use Illuminate\Support\Facades\Storage;
 
 class DosenWaliController extends Controller
 {
@@ -90,71 +88,4 @@ class DosenWaliController extends Controller
         return redirect('/akunDosenWali')->with('success', "Password Akun Dosen Wali dengan NIP $nip Berhasil Direset");
     }
 
-    public function viewProfile(){
-        return view('dosenwali.profile.index');
-    }
-
-    public function editProfile(){
-        return view('dosenwali.profile.edit');
-    }
-
-    public function updateProfile(Request $request){
-        $rules = [
-            'no_telp' => 'required|max:15|regex:/^[0-9]{1,15}$/',
-            'email_sso' => [
-                'required',
-                Rule::unique('dosen_wali')->ignore(auth()->user()->dosen_wali, 'email_sso'),
-                'regex:/^[a-zA-Z0-9._%+-]+@lecturers\.undip\.ac\.id$/i',
-            ],
-        ];
-
-        if($request->file('foto_profil')){
-            $rules['foto_profil'] = 'required|image|mimes:jpeg,png,jpg|max:2048';
-        }
-
-        $errorMassages = [
-            'email_sso.regex' => 'Email SSO harus berakhiran dengan domain lecturers.undip.ac.id',
-        ];
-        
-        $validatedData = $request->validate($rules, $errorMassages);
-
-        $foto = auth()->user()->dosen_wali->foto_profil;
-        if($request->foto_profil){
-            if($foto){
-                Storage::delete($foto);
-            }
-            $validatedData ["foto_profil"] = $request->file('foto_profil')->store('private/profile_photo');
-        }
-
-        DosenWali::where('nip', auth()->user()->dosen_wali->nip)->update($validatedData);
-
-        return redirect('/profile')->with('success', 'Profile berhasil diubah');
-    }
-
-    public function editPassword(){
-        return view('dosenwali.profile.edit_password');
-    }
-
-    public function updatePassword(Request $request){
-        $rules = [
-            'password_lama' => [
-                'required',
-                function ($attribute, $value, $fail) {
-                    if (!Hash::check($value, auth()->user()->password)) {
-                        $fail('Password lama salah');
-                    }
-                },
-            ],
-            'password_baru' => 'required|min:8|max:16|regex:/^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/',
-            'konfirmasi_password' => 'required|same:password_baru',
-        ];
-
-        $validatedData = $request->validate($rules);
-
-        $user = auth()->user();
-        $user->password = Hash::make($validatedData['password_baru']);
-        $user->save(); //syntax highlight error, tapi bisa jalan
-
-        return redirect('/profile')->with('success', 'Password berhasil diubah');
-    }
 }
