+++ app/Http/Controllers/MahasiswaController.php
@@ -8,14 +8,11 @@ use App\Models\User;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Http\Request;
-use Illuminate\Validation\Rule;
 use Maatwebsite\Excel\Facades\Excel;
 use App\Imports\UserImport;
 use App\Imports\MahasiswaImport;
 use App\Exports\MahasiswaExport;
 use App\Models\DosenWali;
-use Illuminate\Support\Facades\Validator;
-use Illuminate\Support\Facades\Storage;
 
 class MahasiswaController extends Controller
 {
@@ -138,74 +135,4 @@ class MahasiswaController extends Controller
         return response()->json(['html' => $view]);
     }
 
-    public function viewProfile(){
-        return view('mahasiswa.profile.index');
-    }
-
-    public function editProfile(){
-        return view('mahasiswa.profile.edit');
-    }
-
-    public function updateProfile(Request $request){
-        $rules = [
-            'no_telp' => 'required|max:15|regex:/^[0-9]{1,15}$/',
-            'email_sso' => [
-                'required',
-                Rule::unique('mahasiswa')->ignore(auth()->user()->mahasiswa, 'email_sso'),
-                'regex:/^[a-zA-Z0-9._%+-]+@students\.undip\.ac\.id$/i',
-            ],
-            'alamat' => 'required|max:255',
-            'kabupaten_kota' => 'required',
-            'provinsi' => 'required',
-        ];
-
-        if($request->file('foto_profil')){
-            $rules['foto_profil'] = 'required|image|mimes:jpeg,png,jpg|max:2048';
-        }
-
-        $errorMassages = [
-            'email_sso.regex' => 'Email SSO harus berakhiran dengan domain students.undip.ac.id',
-        ];
-        
-        $validatedData = $request->validate($rules, $errorMassages);
-
-        $foto = auth()->user()->mahasiswa->foto_profil;
-        if($request->foto_profil){
-            if($foto){
-                Storage::delete($foto);
-            }
-            $validatedData ["foto_profil"] = $request->file('foto_profil')->store('private/profile_photo');
-        }
-
-        Mahasiswa::where('nim', auth()->user()->mahasiswa->nim)->update($validatedData);
-
-        return redirect('/profile')->with('success', 'Profile berhasil diubah');
-    }
-
-    public function editPassword(){
-        return view('mahasiswa.profile.edit_password');
-    }
-
-    public function updatePassword(Request $request){
-        $rules = [
-            'password_lama' => [
-                'required',
-                function ($attribute, $value, $fail) {
-                    if (!Hash::check($value, auth()->user()->password)) {
-                        $fail('Password lama salah');
-                    }
-                },
-            ],
-            'password_baru' => 'required|min:8|max:16|regex:/^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/',
-            'konfirmasi_password' => 'required|same:password_baru',
-        ];
-
-        $validatedData = $request->validate($rules);
-
-        $user = auth()->user();
-        $user->password = Hash::make($validatedData['password_baru']);
-        $user->save(); //syntax highlight error, tapi bisa jalan
-
-        return redirect('/profile')->with('success', 'Password berhasil diubah');
-    }
 }
